+++ src/mofaflex/_core/datasets/__init__.py
@@ -1,3 +1,4 @@
+from .anndatadataset import AnnDataDataset
 from .anndatadictdataset import AnnDataDictDataset
 from .base import MofaFlexDataset, Preprocessor
 from .misc import CovariatesDataset, MofaFlexBatchSampler, StackDataset, merge_covariates