+++ src/mofaflex/tl/_downstream.py
@@ -17,7 +17,7 @@ _logger = logging.getLogger(__name__)
 def test_annotation_significance(
     model: types.MofaFlex | MOFAFLEX,
     annotations: dict[str, pd.DataFrame],
-    data: MuData | dict[str, dict[str, AnnData]] | MofaFlexDataset | None = None,
+    data: MuData | dict[str, dict[str, AnnData]] | AnnData | MofaFlexDataset | None = None,
     corr_adjust: bool = True,
     p_adj_method: str = "fdr_bh",
     min_size: int = 10,